cloud_function/requirements.txt
# Core data processing
pandas>=1.5.0,<2.1.0
numpy>=1.21.0,<2.0.0
openpyxl>=3.1.0,<4.0.0

# HTTP requests and API calls
requests>=2.28.0,<3.0.0

# Configuration and environment
python-dotenv>=0.19.0,<1.0.0
PyYAML>=6.0,<7.0

# PDF generation
fpdf>=1.7.2,<2.0.0

# Google Cloud Platform dependencies
google-cloud-storage>=2.10.0,<3.0.0
google-cloud-functions>=1.8.0,<2.0.0
google-auth>=2.17.0,<3.0.0
google-auth-oauthlib>=1.0.0,<2.0.0
google-auth-httplib2>=0.1.0,<1.0.0
google-api-python-client>=2.95.0,<3.0.0

# Slack integration
slack-sdk>=3.21.0,<4.0.0

# Timezone handling
pytz>=2023.3,<2024.0

# Functions Framework for local testing
functions-framework>=3.4.0,<4.0.0

# Additional utilities
pathlib2>=2.3.7; python_version < "3.4" 